Introduce StateVector (#27)
Wraps the state variables returned to the user in a StateVector, but this is also indexed by the cellnumber, hence according to the docs. The key advantage is that this type allows modifying the underlying data by changing a pointer, allowing for easy update of state variables, and this makes it easier to update the state variables if we want to reuse the buffer for multiple simulations, but with different state variables (typical for rve simulations).